# docs/swagger.yaml
definitions:
...
  models.CreateRoomRequest:
    properties:
      capacity:
        type: integer
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.CreateScheduleRequest:
    properties:
      daysOfWeek:
        items:
          type: integer
        type: array
      endTime:
        type: string
      startTime:
        type: string
    required:
    - daysOfWeek
    - endTime
    - startTime
    type: object
...
  models.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
    type: object
...
  models.ListRoomResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/models.RoomObject'
        type: array
    type: object
...
  models.RoomObject:
    properties:
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.RoomResponse:
    properties:
      room:
        $ref: '#/definitions/models.RoomObject'
    type: object
  models.ScheduleObject:
    properties:
      daysOfWeek:
        items:
          type: integer
        type: array
      endTime:
        type: string
      id:
        type: string
      roomId:
        type: string
      startTime:
        type: string
    type: object
  models.ScheduleResponse:
    properties:
      schedule:
        $ref: '#/definitions/models.ScheduleObject'
    type: object
  models.SlotObject:
    properties:
      end:
        type: string
      id:
        type: string
      roomId:
        type: string
      start:
        type: string
    type: object
  models.SlotsResponse:
    properties:
      slots:
        items:
          $ref: '#/definitions/models.SlotObject'
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Сервис бронирования переговорок
  title: Room Booking Service API
  version: 1.0.0
paths:
...
  /rooms/{roomId}/schedule/create:
    post:
      consumes:
      - application/json
      description: Создание расписания для переговорки (только admin)
      parameters:
      - description: ID переговорки
        in: path
        name: roomId
        required: true
        type: string
      - description: Данные расписания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать расписание
      tags:
      - schedules
  /rooms/{roomId}/slots/list:
    get:
      consumes:
      - application/json
      description: Получить доступные для бронирования слоты по переговорке и дате
      parameters:
      - description: ID переговорки
        in: path
        name: roomId
        required: true
        type: string
      - description: Дата в формате YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SlotsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Доступные слоты
      tags:
      - slots
  /rooms/create:
    post:
      consumes:
      - application/json
      description: Создание новой переговорки (только admin)
      parameters:
      - description: Данные переговорки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать переговорку
      tags:
      - rooms
  /rooms/list:
    get:
      consumes:
      - application/json
      description: Получить список всех переговорок
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListRoomResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список переговорок
      tags:
      - rooms
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"